Rename timestamp publisher for clarity
